requirements.txt
absl-py==1.2.0
astunparse==1.6.3
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
contourpy==1.0.5
cycler==0.11.0
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Login==0.6.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.47.0
h5py==3.7.0
idna==3.3
intel-extension-for-tensorflow==0.0.0.dev1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
matplotlib==3.6.0
numpy==1.23.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
SQLAlchemy==1.4.36
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow_intel==2.10.0
termcolor==1.1.0
threadpoolctl==3.1.0
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
